requirements.txt
pandas
pdfplumber
python-dotenv
gspread
google-api-python-client
google-auth-httplib2
google-auth-oauthlib